LPS-123455 Convert the deleteDDLRecord action method to a mvcActionCommand and apply to usages
