+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-forms-web/src/main/java/com/liferay/portal/workflow/kaleo/forms/web/internal/portlet/KaleoFormsAdminPortlet.java
@@ -168,70 +168,6 @@ import org.osgi.service.component.annotations.Reference;
 )
 public class KaleoFormsAdminPortlet extends MVCPortlet {
 
-
-	/**
-	 * Deletes the <code>DDLRecord</code> (in the
-	 * <code>com.liferay.dynamic.data.lists.api</code> module) associated with
-	 * the record IDs from the action request. This method also deletes the
-	 * <code>WorkflowInstanceLink</code>s (in
-	 * <code>com.liferay.portal.kernel</code>) associated with each record ID
-	 * and matching name of the <code>KaleoProcess</code> implementation (in the
-	 * <code>com.liferay.portal.workflow.kaleo.forms.api</code> module). This
-	 * method uses <code>TransactionInvokerUtil</code> (in
-	 * <code>com.liferay.portal.kernel</code>) to ensure all tasks are performed
-	 * in a single transaction.
-	 *
-	 * @param  actionRequest the request from which to get the request
-	 *         parameters
-	 * @param  actionResponse the response to receive the render parameters
-	 * @throws Exception if an exception occurred
-	 */
-	public void deleteDDLRecord(
-			ActionRequest actionRequest, ActionResponse actionResponse)
-		throws Exception {
-
-		ServiceContext serviceContext = ServiceContextFactory.getInstance(
-			DDLRecord.class.getName(),
-			_portal.getUploadPortletRequest(actionRequest));
-
-		checkKaleoProcessPermission(serviceContext, ActionKeys.DELETE);
-
-		final ThemeDisplay themeDisplay =
-			(ThemeDisplay)actionRequest.getAttribute(WebKeys.THEME_DISPLAY);
-
-		long[] ddlRecordIds = getDDLRecordIds(actionRequest);
-
-		for (final long ddlRecordId : ddlRecordIds) {
-			try {
-				Callable<Void> callable = new Callable<Void>() {
-
-					@Override
-					public Void call() throws Exception {
-						_ddlRecordLocalService.deleteRecord(ddlRecordId);
-
-						_workflowInstanceLinkLocalService.
-							deleteWorkflowInstanceLinks(
-								themeDisplay.getCompanyId(),
-								themeDisplay.getScopeGroupId(),
-								KaleoProcess.class.getName(), ddlRecordId);
-
-						return null;
-					}
-
-				};
-
-				TransactionInvokerUtil.invoke(_transactionConfig, callable);
-			}
-			catch (Throwable throwable) {
-				if (throwable instanceof PortalException) {
-					throw (PortalException)throwable;
-				}
-
-				throw new SystemException(throwable);
-			}
-		}
-	}
-
 	/**
 	 * Deletes the <code>KaleoDraftDefinition</code> (in the
 	 * <code>com.liferay.portal.workflow.kaleo.designer.api</code> module) by
@@ -501,24 +437,6 @@ public class KaleoFormsAdminPortlet extends MVCPortlet {
 		else {
 			super.doDispatch(renderRequest, renderResponse);
 		}
-	}
-
-	/**
-	 * Returns an array of the DDL record IDs obtained from the action request.
-	 *
-	 * @param  actionRequest the request from which to get the request
-	 *         parameters
-	 * @return an array of the DDL record IDs
-	 */
-	protected long[] getDDLRecordIds(ActionRequest actionRequest) {
-		long ddlRecordId = ParamUtil.getLong(actionRequest, "ddlRecordId");
-
-		if (ddlRecordId > 0) {
-			return new long[] {ddlRecordId};
-		}
-
-		return StringUtil.split(
-			ParamUtil.getString(actionRequest, "ddlRecordIds"), 0L);
 	}
 
 	/**
@@ -828,13 +746,6 @@ public class KaleoFormsAdminPortlet extends MVCPortlet {
 
 	@Reference
 	protected StorageEngine storageEngine;
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		KaleoFormsAdminPortlet.class);
-
-	private static final TransactionConfig _transactionConfig =
-		TransactionConfig.Factory.create(
-			Propagation.REQUIRES_NEW, new Class<?>[] {Exception.class});
 
 	@Reference
 	private DDLExporterFactory _ddlExporterFactory;
