+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-forms-web/src/main/resources/META-INF/resources/admin/view_kaleo_process.jsp
@@ -210,7 +210,7 @@ portletURL.setParameter("kaleoProcessId", String.valueOf(kaleoProcess.getKaleoPr
 							),
 						},
 
-						<portlet:actionURL name="deleteDDLRecord" var="deleteDDLRecordURL">
+						<portlet:actionURL name="/kaleo_forms/delete_record" var="deleteDDLRecordURL">
 							<portlet:param name="mvcPath" value="/admin/view_kaleo_process.jsp" />
 							<portlet:param name="redirect" value="<%= currentURL %>" />
 							<portlet:param name="kaleoProcessId" value="<%= String.valueOf(kaleoProcess.getKaleoProcessId()) %>" />
